# 1. GestIC(Aries)/backend/src/config/swagger.yaml
swagger: "2.0"
info:
  description: "GestIC App API documentation"
  version: "1.0.0"
  title: "GestIC API"
host: "localhost:3333"
...
paths:
...
  /project:
    post:
      tags:
      - "Project"
      summary: "Insert a project"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              description:
                type: "string"
              type:
                type: "string"
      responses:
        "200":
          description: "Inserted a project"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              type:
                type: "string"
        "500":
          description: "Could not insert a project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    put:
      tags:
      - "Project"
      summary: " Update project information"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          description: "Any quantity of the following parameters can be sent, but id is required."
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              newOwner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              type:
                type: "string"
      responses:
        "200":
          description: "Sucessfully updated a project"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              type:
                type: "string"
        "500":
          description: "Could not update a project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /project/:id:
    get:
      tags:
        - "Project"
      summary: "Get project information"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Obtained a project"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              type:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not obtain a project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Project"
      summary: "Delete project"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Successfully removed a project"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              type:
                type: "string"
        "500":
          description: "Could not remove a project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /project/members:
    post:
      tags:
        - "Project"
      summary: "Add a member to project"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
      responses:
        "200":
          description: "Added a member"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              type:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not add a member to project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Project"
      summary: "Remove a member from project"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
      responses:
        "200":
          description: "Removed a member"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              type:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not remove a member from project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
...
  /informative:
    get:
      tags:
      - "Informatives"
      summary: "Get a list of all informatives"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Obtained a list of all informatives"
          schema:
            type: "array"
            produces: "application/json"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                  format: "uuid"
                owner:
                  type: "string"
                  format: "uuid"
                title:
                  type: "string"
                content:
                  type: "string"
        "500":
          description: "Could not obtain a list of all informatives"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    post:
      tags:
      - "Informatives"
      summary: "Insert a informative"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              content:
                type: "string"
      responses:
        "200":
          description: "Inserted a informative"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
        "500":
          description: "Could not insert a informative"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    put:
      tags:
      - "Informatives"
      summary: " Update informative information"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          description: "Any quantity of the following parameters can be sent, but id is required."
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
      responses:
        "200":
          description: "Sucessfully updated a informative"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
        "500":
          description: "Could not update a informative"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /informative/:id:
    get:
      tags:
        - "Informatives"
      summary: "Get informative information"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Obtained a informative"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
        "500":
          description: "Could not obtain a informative"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Informatives"
      summary: "Delete informative"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Successfully removed a informative"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              owner:
                type: "string"
                format: "uuid"
              title:
                type: "string"
              content:
                type: "string"
        "500":
          description: "Could not remove a informative"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
...